extract wallet code to new module
